Move openeit back to OpenEIT
In agreement with @sebastianriese:
```
While this is in violation of PEP8 it is a conscious violation of PEP8 (which PEP8 explicitly supports).

My reasoning is that (a) it is in concordance with many other packages, e.g. OpenSSL, OpenGL, etc. of the form OpenTLA (b) openeit is hard to read and underscores are discouraged in package names.
```